Refactor FreeCell solver: modularize, enhance logging
- Move board dump logic to FreeCellGameBase for reuse and accuracy
- Extract FreeCellMove and FreeCellSolver into separate files
- Expand FreeCellMove with tracing and ApplyMove support
- Improve solver logic and prevent undo-reverse moves
- Update tests to use new dump method and disable auto-foundation
- Add placeholder for auto-solve test
- Minor code cleanup and sequence length calculation improvements
