+++ TestProject1/FreeCellSolverTests.cs
@@ -4,12 +4,11 @@ using Microsoft.Playwright;
 using System.Text.Json;
 using System.Diagnostics;
 using static Microsoft.Playwright.Assertions;
-using Client.Games.Cards.Models;
 
 namespace TestProject1
 {
     [TestClass]
-    public class FreeCellSolverTests : InteractiveTestBase
+    public partial class FreeCellSolverTests : InteractiveTestBase
     {
         [ClassInitialize]
         public static async Task ClassInitialize(TestContext context)
@@ -285,7 +284,7 @@ for (int i = 0; i < colCount; i++)
             await mover.MoveTableauToFoundationAsync(columnIndex: 4, foundationIndex: 0);
             await mover.MoveTableauToFreeCellAsync(columnIndex: 4, freeCellIndex: 3);
             await mover.MoveFreeCellToTableauAsync(freeCellIndex: 3, columnIndex: 2);
-            mover.dumpAllToLog("After some moves");
+            LogAction(mover.gameService.dumpAllToLog($"After {mover.gameService.MoveCount} moves"));
             /*
                 FreeCells:   8♥   3♣   3♥      
             Tableau:
@@ -362,7 +361,7 @@ for (int i = 0; i < colCount; i++)
                 }
                 else
                 {
-                    mover.dumpAllToLog($"No moves found by solver at move count {mover.gameService.MoveCount}.");
+                    LogAction(mover.gameService.dumpAllToLog($"No moves found by solver at move count {mover.gameService.MoveCount}."));
                     Assert.Fail();
                     break; // no moves found, should not happen unless we have a bug in FindMoves
                 }
@@ -373,197 +372,56 @@ for (int i = 0; i < colCount; i++)
             await Task.WhenAny(Task.Delay(1500000), pageClosedTcs.Task); // Wait for either the page to close or a timeout)
 
         }
-        public class FreeCellMove
+        [TestMethod]
+        [TestCategory("Manual")]
+        public async Task AutoSolve_FindSolution()
         {
-            /*
-             * A move can be between any combination of Foundation/Freecell/Tableau
-             * If tableau to tableau, source and target index are column indexes and cardCount is how many cards from the bottom of the source column
-             * The Seq count for a board is the sum of all sequences that are valid (consecutive, altering red/black). Increasing overall seq count is good
-             * Some moves may move a card to the foundation or freecell, which reduces seq count
-             */
-            public SourceType sourceType { get; set; }
-            public SourceType targetType { get; set; }
-            public int sourceIndex { get; set; } // column # or index of FreeCell or Foundation
-            public int targetIndex { get; set; } // column # or index of FreeCell or Foundation
-            public int srcColumnIndex { get; set; } // only for tableau to tableau moves
-            public int cardCount { get; set; } // only for tableau to tableau moves, how many cards from the bottom of the source column
+            var gameId = 12345;
+            var gameService = new FreeCellGameService();
+            gameService.InitializeGame(gameId);
+            var moveHistory = new List<FreeCellMove>();
+            var solver = await FreeCellSolver.CreateAsync(gameService, moveHistory);
+            LogAction(gameService.dumpAllToLog($"Initial layout game {gameService.GameId}"));
 
-            public int score { get; set; }
-            public int deltaSequenceCount { get; set; }
-            public Card CardMoved { get; set; }
-            public FreeCellMove(Card cardMoved)
-            {
-                CardMoved = cardMoved;
-            }
-            public override string ToString() => $"{sourceType}[{sourceIndex}] -> {targetType}[{targetIndex}] (cards: {cardCount}, score: {score})";
-        }
-        public class FreeCellSolver
-        {
-            private FreeCellGameService _gameService; // current state of board including undo
-            private FreeCellGameBase _gameClone; // state of board as we manipulate it
-            private List<FreeCellMove> _moveHistory; // so we don't repeat moves that we just did
 
-            public FreeCellSolver(FreeCellGameService gameService, List<FreeCellMove> moveHistory)
-            {
-                _gameService = gameService;
-                _gameClone = gameService.Clone();
-                _gameClone.AutoMoveToFoundationDisable = true;
-                _moveHistory = moveHistory;
-            }
+            //var pageClosedTcs = new TaskCompletionSource<bool>();
+            //var page = await GetPageForGame(gameId, pageClosedTcs);
 
-            public static async Task<FreeCellSolver> CreateAsync(FreeCellGameService freeCellGameService, List<FreeCellMove> moveHistory)
-            {
-                var solver = new FreeCellSolver(freeCellGameService, moveHistory);
-                return solver;
-            }
-            public List<FreeCellMove> FindMoves()
-            {
-                var lstMoves = new List<FreeCellMove>();
-                int nFreeCells = _gameClone.EmptyFreeCellCount;
-                var bottomSequences = _gameClone.GetBottomSequenceLengths();
-                var sumSeqLen = bottomSequences.Sum(); // sum of all sequence lengths from each column. A good move will often increase this by creating longer sequences, a bad move will decrease it by breaking sequences up
-                var maxScoreSoFar = 0;
-                bool AddNewMove(FreeCellMove move)
-                {
-                    var didit = false;
-                    if (!moveWouldJustUndoPriorMove(move))
-                    {
-                        if (move.score > maxScoreSoFar)
-                        {
-                            maxScoreSoFar = move.score;
-                        }
-                        lstMoves.Add(move);
-                        didit = true;
-                    }
-                    return didit;
-                }
-                // first see if any of the freecells can be moved to a foundation or tableau
-                for (int i = 0; i < _gameClone.FreeCells.Count; i++)
-                {
-                    var freecellCard = _gameClone.FreeCells[i];
-                    if (freecellCard == null) continue;
-                    // Check if we can move this card to a foundation
-                    var foundationIndex = _gameClone.CanMoveToAnyFoundation(freecellCard);
-                    if (foundationIndex >= 0)
-                    {
-                        AddNewMove(new FreeCellMove(freecellCard)
-                        {
-                            sourceType = SourceType.FreeCell,
-                            targetType = SourceType.Foundation,
-                            sourceIndex = i,
-                            targetIndex = foundationIndex,
-                            cardCount = 1,
-                            score = 20 // arbitrary score for now
-                        });
-                        //return lstMoves; // prioritize moving to foundation
-                    }
-                    for (var dstCol = 0; dstCol < _gameClone.Tableau.Count; dstCol++)
-                    {
-                        if (_gameClone.CanMoveFreeCellToTableau(i, dstCol))
-                        {
-                            // if the dest column is empty, don't do it: no gain in moving free card to empty column
-                            if (_gameClone.Tableau[dstCol].Count > 0)
-                            {
-                                AddNewMove(new FreeCellMove(freecellCard)
-                                {
-                                    sourceType = SourceType.FreeCell,
-                                    targetType = SourceType.Tableau,
-                                    sourceIndex = i,
-                                    targetIndex = dstCol,
-                                    cardCount = 1,
-                                    score = 15 // arbitrary score for now
-                                });
-                            }
-                        }
-                    }
-                }
+            //var mover = await FreeCellMover.CreateAsync(page, InteractiveTestBase._IsDebugging);
+            //var moveHistory = new List<FreeCellMove>();
+            ////await mover.MoveTableauToTableauAsync(srcColumnIndex: 7, destColumnIndex: 4, cardCount: 1);
 
-                for (int srcCol = 0; srcCol < _gameClone.Tableau.Count; srcCol++)
-                {
-                    var column = _gameClone.Tableau[srcCol];
-                    if (column.Count == 0) continue;
-                    var seqlen = _gameClone.GetBottomSequenceLength(srcCol);
-                    var topCard = column[^seqlen];
-                    var botCard = column[^1];
-                    //Log($"Col {i} has {column.Count}, bottom seq len={seqlen}, Top= {topCard} Bot={botCard}");
-                    // Check if we can move this card to a foundation
-                    var foundationIdx = _gameClone.CanMoveToAnyFoundation(botCard);
-                    if (foundationIdx >= 0)
-                    {
-                        //Log($"Can move {topCard} from column {i} to foundation");
-                        AddNewMove(new FreeCellMove(botCard)
-                        {
-                            sourceType = SourceType.Tableau,
-                            targetType = SourceType.Foundation,
-                            sourceIndex = srcCol,
-                            targetIndex = foundationIdx,
-                            cardCount = 1,
-                            score = 10 // arbitrary score for now
-                        });
-                        //return lstMoves;
-                    }
-                    for (var dstCol = 0; dstCol < _gameClone.Tableau.Count; dstCol++)
-                    {
-                        if (srcCol == dstCol) continue;
-                        // if the destination column is empty, and the seqlen is the entire column, don't do anything. Moving an entire column is a no-op
-                        if (_gameClone.Tableau[dstCol].Count == 0 && seqlen == column.Count)
-                        {
-                            continue;
-                        }
-                        int maxMovable = _gameClone.CalculateMaxMovableCards(SourceType.Tableau, dstCol);
-                        if (seqlen > maxMovable)
-                        {
-                            continue;
-                        }
-                        if (_gameClone.CanMoveTableauToTableau(srcCol, dstCol, seqlen))
-                        {
-                            //Log($"Can move {seqlen} cards from column {i} to column {j}");
-                            AddNewMove(new FreeCellMove(topCard)
-                            {
-                                sourceType = SourceType.Tableau,
-                                targetType = SourceType.Tableau,
-                                sourceIndex = srcCol,
-                                targetIndex = dstCol,
-                                cardCount = seqlen,
-                                score = 5 + seqlen // arbitrary scoring that favors longer moves
-                            });
-                        }
-                    }
+            //while (!mover.gameService.IsGameWon)
+            //{
+            //    if (mover.gameService.MoveCount >= 22)
+            //    {
+            //        "bpt".ToString();
+            //        //mover.dumpAllToLog();
 
-                    // Check if we can move this card to a free cell
-                    if (nFreeCells >= seqlen && maxScoreSoFar < 2) // don't move to freecell if we can move to foundation or tableau to tableau
-                    {
-                        AddNewMove(new FreeCellMove(botCard)
-                        {
-                            sourceType = SourceType.Tableau,
-                            targetType = SourceType.FreeCell,
-                            sourceIndex = srcCol,
-                            targetIndex = _gameClone.FindAnyFreeCell(),
-                            cardCount = 1,
-                            score = 1 // arbitrary score for now
-                        });
-                        //Console.WriteLine($"Can move {topCard} from tableau column {i + 1} to a free cell");
-                    }
-                }
-                return lstMoves;
-            }
+            //    }
+            //    var solver = await FreeCellSolver.CreateAsync(mover.gameService, moveHistory);
+            //    var moves = solver.FindMoves();
+            //    if (moves.Count > 0)
+            //    {
+            //        var bestMoves = moves.OrderByDescending(m => m.score);
+            //        var bestMove = bestMoves.First();
+            //        //do the best move
+            //        Log($"{mover.gameService.MoveCount,2} {bestMove.CardMoved} {bestMove.sourceType} {bestMove.sourceIndex} -> {bestMove.targetType} {bestMove.targetIndex}, cnt={bestMove.cardCount}, score={bestMove.score} #moves = {moves.Count}");
+            //        await mover.doMoveAsync(bestMove);
+            //        moveHistory.Add(bestMove);
+            //    }
+            //    else
+            //    {
+            //        mover.dumpAllToLog($"No moves found by solver at move count {mover.gameService.MoveCount}.");
+            //        Assert.Fail();
+            //        break; // no moves found, should not happen unless we have a bug in FindMoves
+            //    }
 
-            private bool moveWouldJustUndoPriorMove(FreeCellMove newMove)
-            {
-                if (_moveHistory.Count == 0) return false;
-                var lastMove = _moveHistory[^1];
-                // if the new move is the exact opposite of the last move, then it would just undo it
-                if (newMove.sourceType == lastMove.targetType &&
-                    newMove.targetType == lastMove.sourceType &&
-                    newMove.sourceIndex == lastMove.targetIndex &&
-                    newMove.targetIndex == lastMove.sourceIndex &&
-                    //newMove.CardMoved == lastMove.CardMoved &&
-                    newMove.cardCount == lastMove.cardCount)
-                {
-                    return true;
-                }
-                return false;
-            }
+            //}
+            ////await Task.Delay(5000);
+            ////pageClosedTcs.TrySetResult(true);
+            //await Task.WhenAny(Task.Delay(1500000), pageClosedTcs.Task); // Wait for either the page to close or a timeout)
+
         }
     }
 }
